[FIX] vlib:  Use pow(), not powf(), for OS X 10.2
That we didn't make this change previously, when switching from float to
double, was an oversight.  In any case, Mac OS X "Jaguar" has no powf().

+++ v/vlib/vlib/types/float.cc
@@ -232,7 +232,7 @@ namespace vlib
 			case Op_multiply:  a *= b;  break;
 			case Op_divide:    a /= b;  break;
 			
-			case Op_empower:   a = powf( a, b );  break;
+			case Op_empower:   a = pow( a, b );  break;
 			
 			default:
 				// not reached
